tax files for callahan fetcher

+++ bin/fetcher_callahan_old.py
@@ -35,7 +35,8 @@ if __name__ == '__main__':
     os.makedirs(data_folder, exist_ok=True)
 
     for f in os.listdir(data_folder):
-        os.remove(f'{data_folder}/{f}')
+        if '.pdf' in f:
+            os.remove(f'{data_folder}/{f}')
 
     with tempfile.TemporaryFile() as tfile:
         response = requests.get(url_tax)
